Rename lib validation files
